sass_system/requirements.txt
amqp==5.3.1
asgiref==3.11.1
billiard==4.2.4
celery==5.6.2
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
django==4.2.29
django-cors-headers==4.9.0
django-filter==25.1
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
drf-yasg==1.21.15
exceptiongroup==1.3.1
inflection==0.5.1
kombu==5.6.2
packaging==26.0
prompt-toolkit==3.0.52
psycopg2-binary==2.9.11
pyjwt==2.12.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2026.1.post1
pyyaml==6.0.3
six==1.17.0
sqlparse==0.5.5
typing-extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
uritemplate==4.2.0
vine==5.1.0
wcwidth==0.6.0
